move up and down

+++ sources/movement/location.h
@@ -15,8 +15,8 @@
 #include <stdint.h>
 
 typedef struct Location {
-    int8_t x;
-    int8_t y;
+    int16_t x;
+    int16_t y;
     int8_t zoneId;
 } Location;
